refactor(kfx): align asyncio runtime with current dev
Consolidate the service bootstrap at the existing Python host boundary, refresh rooted KFX and KFD evidence, and retain fail-closed relay identity validation on the latest protected dev baseline.
