+++ docs/adr/KF-ADR-019fb64f-ba63-7620-a384-063adec7af2f.md
@@ -3,8 +3,9 @@ metadata_schema: kungfu.document-metadata/v1
 doc_type: architecture-decision
 adr_id: KF-ADR-019fb64f-ba63-7620-a384-063adec7af2f
 decision_status: accepted
-implementation_status: staged
-implementation_prs: []
+implementation_status: implemented
+implementation_prs: [https://github.com/kungfu-systems/kungfu/pull/2021]
+closure_pr: https://github.com/kungfu-systems/kungfu/pull/2021
 qualification_refs: [docs/qualification/python-kfx-asyncio-runtime.contract.json, framework/core/tests/python/test_event_loop_concurrency.py, framework/core/tests/python/test_kfx_async_capability.py, framework/core/tests/python/test_kfx_python_service.py, framework/tui/src/service-host-python.test.ts]
 review_state: self-reviewed
 sensitivity: public
@@ -19,7 +20,7 @@ ai_provenance: GPT-5 via Codex on 2026-07-31; based on repository sources, nativ
 
 # KF-ADR-019fb64f-ba63-7620-a384-063adec7af2f: Python KFX services use standard asyncio behind a journal bridge
 
-- Status: accepted
+- Status: accepted; implementation delivered by PR #2021
 - Date: 2026-07-31
 
 ## Context
