+++ docs/architecture/adr-map.md
@@ -170,7 +170,7 @@ review`; evidence is the number of declared qualification references.
 | KF · e8ab | [KFX admission consumes KFD facts and exact Buildchain attestations](../adr/KF-ADR-019f86da-4f90-7d6c-926a-ddd27dbde8ab.md) | kfd-aware-kfx-trust-buildchain-admission | accepted / implemented / self-reviewed | 7 | [KF · 8708](../adr/KF-ADR-019f86da-4f90-79d7-a4b7-044fcf998708.md) (shared terms: kfx)<br>[KF · 16be](../adr/KF-ADR-019f86da-4f90-7afa-a1e1-0510f00916be.md) (shared terms: kfx)<br>[KF · 31c6](../adr/KF-ADR-019f86da-4f90-7d41-a4a0-e6b01d4b31c6.md) (shared terms: kfx)<br>[KF · 8bcc](../adr/KF-ADR-019f86da-4f90-73f3-b027-c343b2bc8bcc.md) (shared terms: kfx) |
 | KF · 6704 | [control axis — the Node watcher snapshot model](../adr/KF-ADR-019f86da-4f90-7fb3-a803-393d3bbe6704.md) | — | withdrawn / not-applicable / self-reviewed | 0 | [KF · 053d](../adr/KF-ADR-019f86da-4f90-7a30-8697-5c648120053d.md) (shared terms: axis, control)<br>[KF · d390](../adr/KF-ADR-019f86da-4f90-7809-898a-ccc0f52bd390.md) (shared terms: node) |
 | KF · b500 | [Layered APIs share one C ABI waist and each identity protocol owns its canonical bytes](../adr/KF-ADR-019f87cb-b4e7-7cda-80ec-be5aceb7b500.md) | layered-api-and-protocol-owned-canonical-encoding | accepted / staged / self-reviewed | 13 | [KF · 8f82](../adr/KF-ADR-019f86da-4f90-74d7-9ddd-84adc0f38f82.md) (shared terms: one, share)<br>[KF · 8bcc](../adr/KF-ADR-019f86da-4f90-73f3-b027-c343b2bc8bcc.md) (shared terms: owns)<br>[KF · d390](../adr/KF-ADR-019f86da-4f90-7809-898a-ccc0f52bd390.md) (shared terms: one)<br>[KF · ef05](../adr/KF-ADR-019f86da-4f90-73ff-9543-f0a4f0beef05.md) (shared terms: layered) |
-| KF · af2f | [Python KFX services use standard asyncio behind a journal bridge](../adr/KF-ADR-019fb64f-ba63-7620-a384-063adec7af2f.md) | python-kfx-asyncio-runtime | accepted / staged / self-reviewed | 5 | [KF · 8708](../adr/KF-ADR-019f86da-4f90-79d7-a4b7-044fcf998708.md) (shared terms: kfx)<br>[KF · 3f06](../adr/KF-ADR-019f86da-4f90-7a4b-b1a3-256ce6fa3f06.md) (shared terms: python)<br>[KF · 053d](../adr/KF-ADR-019f86da-4f90-7a30-8697-5c648120053d.md) (shared terms: python)<br>[KF · 16be](../adr/KF-ADR-019f86da-4f90-7afa-a1e1-0510f00916be.md) (shared terms: kfx) |
+| KF · af2f | [Python KFX services use standard asyncio behind a journal bridge](../adr/KF-ADR-019fb64f-ba63-7620-a384-063adec7af2f.md) | python-kfx-asyncio-runtime | accepted / implemented / self-reviewed | 5 | [KF · 8708](../adr/KF-ADR-019f86da-4f90-79d7-a4b7-044fcf998708.md) (shared terms: kfx)<br>[KF · 3f06](../adr/KF-ADR-019f86da-4f90-7a4b-b1a3-256ce6fa3f06.md) (shared terms: python)<br>[KF · 053d](../adr/KF-ADR-019f86da-4f90-7a30-8697-5c648120053d.md) (shared terms: python)<br>[KF · 16be](../adr/KF-ADR-019f86da-4f90-7afa-a1e1-0510f00916be.md) (shared terms: kfx) |
 
 </details>
